Encapsulating the database interactions with the Player data into a DAO object.
Orgainsed all database interactions within the new DAO and removed unnecessary additional calls from Unit Tests.

Tidy-up connection management to ensure correct resource release.
