+++ src/main/java/com/ffxivcensus/gatherer/GathererController.java
@@ -3,14 +3,12 @@ package com.ffxivcensus.gatherer;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-import org.apache.commons.lang.BooleanUtils;
-import org.apache.commons.lang.StringUtils;
 import com.ffxivcensus.gatherer.config.ApplicationConfig;
-import com.ffxivcensus.gatherer.config.ConfigurationBuilder;
+import com.ffxivcensus.gatherer.player.PlayerBeanDAO;
+import com.ffxivcensus.gatherer.player.PlayerBuilder;
 
 /**
  * GathererController class of character gathering program. This class makes calls to fetch records from the lodestone, and then
@@ -64,12 +62,14 @@ public class GathererController {
         } else { // Else configured correctly
             try {
 
+                PlayerBeanDAO dao = new PlayerBeanDAO(appConfig);
+
                 if(appConfig.isSplitTables()) { // If specified to split tables
                     for(String realm : realms) { // Create a table for each realm
-                        createTable("tbl" + realm.toLowerCase() + appConfig.getTableSuffix());
+                        dao.createTable("tbl" + realm.toLowerCase() + appConfig.getTableSuffix());
                     }
                 } else { // Else just create a single table
-                    createTable(appConfig.getTableName());
+                    dao.createTable(appConfig.getTableName());
                 }
 
                 if(appConfig.getStartId() > appConfig.getEndId()) {
@@ -132,60 +132,6 @@ public class GathererController {
     }
 
     /**
-     * Create the table to write character records to.
-     *
-     * @param tableName the name of the table to setup.
-     */
-    private void createTable(final String tableName) {
-        // Create DB table if it doesn't exist
-        // Open connection
-        Connection conn = openConnection();
-        try {
-            Statement st = conn.createStatement();
-            StringBuilder sbSQL = new StringBuilder();
-            sbSQL.append("CREATE TABLE IF NOT EXISTS ");
-            sbSQL.append(tableName);
-            sbSQL.append(" (id INTEGER PRIMARY KEY,name TEXT,realm TEXT,race TEXT,gender TEXT,grand_company TEXT,free_company TEXT,");
-            sbSQL.append("level_gladiator INTEGER,level_pugilist INTEGER,level_marauder INTEGER,level_lancer INTEGER,level_archer INTEGER,");
-            sbSQL.append("level_rogue INTEGER,level_conjurer INTEGER,level_thaumaturge INTEGER,level_arcanist INTEGER,level_darkknight INTEGER, level_machinist INTEGER,");
-            sbSQL.append("level_astrologian INTEGER, level_scholar INTEGER, level_redmage INTEGER, level_samurai INTEGER, level_carpenter INTEGER, level_blacksmith INTEGER,");
-            sbSQL.append("level_armorer INTEGER,level_goldsmith INTEGER,level_leatherworker INTEGER,level_weaver INTEGER,level_alchemist INTEGER,level_culinarian INTEGER,");
-            sbSQL.append("level_miner INTEGER,level_botanist INTEGER,level_fisher INTEGER");
-            if(appConfig.isStoreProgression()) {
-                sbSQL.append(",");
-                sbSQL.append("p30days BIT, p60days BIT, p90days BIT, p180days BIT, p270days BIT,p360days BIT,p450days BIT,p630days BIT,p960days BIT,");
-                sbSQL.append("prearr BIT, prehw BIT, presb BIT, arrartbook BIT, hwartbookone BIT, hwartbooktwo BIT, hasencyclopedia BIT, ");
-                sbSQL.append("beforemeteor BIT, beforethefall BIT, ps4collectors BIT, ");
-                sbSQL.append("soundtrack BIT,saweternalbond BIT,sightseeing BIT,comm50 BIT,moogleplush BIT,");
-                sbSQL.append("topazcarubuncleplush BIT,emeraldcarbuncleplush BIT,");
-                sbSQL.append("hildibrand BIT, dideternalbond BIT, arrcollector BIT,");
-                sbSQL.append("kobold BIT, sahagin BIT, amaljaa BIT, sylph BIT,  moogle BIT, vanuvanu BIT, vath BIT,");
-                sbSQL.append("arr_25_complete BIT,hw_complete BIT, hw_31_complete BIT, hw_33_complete BIT, legacy_player BIT");
-            }
-            if(appConfig.isStoreMounts()) {
-                sbSQL.append(",mounts TEXT");
-            }
-            if(appConfig.isStoreMinions()) {
-                sbSQL.append(",minions TEXT");
-            }
-            if(appConfig.isStoreActiveDate()) {
-                sbSQL.append(",");
-                sbSQL.append("date_active DATE");
-            }
-            if(appConfig.isStorePlayerActive()) {
-                sbSQL.append(",");
-                sbSQL.append("is_active BIT");
-            }
-            sbSQL.append(");");
-
-            st.executeUpdate(sbSQL.toString());
-        } catch(SQLException e) {
-            e.printStackTrace();
-        }
-        closeConnection(conn);
-    }
-
-    /**
      * Method to gather data for characters in specified range.
      */
     private void gatherRange() {
@@ -196,7 +142,7 @@ public class GathererController {
         ExecutorService executor = Executors.newFixedThreadPool(appConfig.getThreadLimit());
 
         while(nextID <= appConfig.getEndId()) {
-            Gatherer worker = new Gatherer(this, nextID);
+            Gatherer worker = new Gatherer(this, new PlayerBeanDAO(appConfig), nextID);
             executor.execute(worker);
 
             nextID++;
@@ -206,142 +152,6 @@ public class GathererController {
         while(!executor.isTerminated()) {
             // Wait patiently for the executor to finish off everything submitted
         }
-    }
-
-    /**
-     * Write a player record to database
-     */
-    protected String writeToDB(final PlayerBean player) {
-        String strOut = "";
-        // Open connection
-        Connection conn = openConnection();
-        try {
-            Statement st = conn.createStatement();
-
-            // Declare string builders to build up components of statement
-            StringBuilder sbFields = new StringBuilder();
-            StringBuilder sbValues = new StringBuilder();
-
-            // Set default table name
-            String tableName;
-            // Determine table to write to
-            if(appConfig.isSplitTables()) {
-                tableName = "tbl" + player.getRealm().toLowerCase() + appConfig.getTableSuffix();
-            } else {
-                tableName = this.appConfig.getTableName() + appConfig.getTableSuffix();
-            }
-
-            sbFields.append("INSERT IGNORE INTO ").append(tableName).append(" (");
-            sbValues.append(" VALUES (");
-
-            sbFields.append("id, name, realm, race, gender, grand_company,free_company,");
-            sbValues.append(player.getId() + ",\"" + player.getPlayerName() + "\",\"" + player.getRealm() + "\",\"" + player.getRace()
-                            + "\",'" + player.getGender() + "','" + player.getGrandCompany() + "',\"" + player.getFreeCompany() + "\",");
-
-            sbFields.append("level_gladiator, level_pugilist, level_marauder,level_lancer, level_archer, level_rogue,");
-            sbValues.append(player.getLvlGladiator() + "," + player.getLvlPugilist() + "," + player.getLvlMarauder() + ","
-                            + player.getLvlLancer() + "," + player.getLvlArcher() + "," + player.getLvlRogue() + ",");
-
-            sbFields.append("level_conjurer, level_thaumaturge, level_arcanist, level_astrologian, level_darkknight,"
-                            + " level_machinist,");
-            sbValues.append(player.getLvlConjurer() + "," + player.getLvlThaumaturge() + "," + player.getLvlArcanist() + ","
-                            + player.getLvlAstrologian() + "," + player.getLvlDarkKnight() + "," + player.getLvlMachinist() + ",");
-
-            sbFields.append("level_scholar, level_redmage, level_samurai,");
-            sbValues.append(player.getLvlScholar() + "," + player.getLvlRedMage() + "," + player.getLvlSamurai() + ",");
-
-            sbFields.append("level_carpenter, level_blacksmith, level_armorer, level_goldsmith, level_leatherworker, level_weaver, level_alchemist,");
-            sbValues.append(player.getLvlCarpenter() + "," + player.getLvlBlacksmith() + "," + player.getLvlArmorer() + ","
-                            + player.getLvlGoldsmith() + "," + player.getLvlLeatherworker() + "," + player.getLvlWeaver() + ","
-                            + player.getLvlAlchemist() + ",");
-
-            sbFields.append("level_culinarian, level_miner, level_botanist, level_fisher");
-            sbValues.append(player.getLvlCulinarian() + "," + player.getLvlMiner() + "," + player.getLvlBotanist() + ","
-                            + player.getLvlFisher());
-
-            if(appConfig.isStoreProgression()) {
-                sbFields.append(",");
-                sbValues.append(",");
-
-                sbFields.append("p30days, p60days, p90days, p180days, p270days, p360days, p450days, p630days, p960days,");
-                sbValues.append(booleanToInt(player.isHas30DaysSub()) + "," + booleanToInt(player.isHas60DaysSub()) + ","
-                                + booleanToInt(player.isHas90DaysSub()) + "," + booleanToInt(player.isHas180DaysSub()) + ","
-                                + booleanToInt(player.isHas270DaysSub()) + "," + booleanToInt(player.isHas360DaysSub()) + ","
-                                + booleanToInt(player.isHas450DaysSub()) + "," + booleanToInt(player.isHas630DaysSub()) + ","
-                                + booleanToInt(player.isHas960DaysSub()) + ",");
-
-                sbFields.append("prearr, prehw, presb, arrartbook, hwartbookone, hwartbooktwo, hasencyclopedia, beforemeteor, beforethefall, soundtrack, saweternalbond, "
-                                + "sightseeing, arr_25_complete, comm50, moogleplush, topazcarubuncleplush, emeraldcarbuncleplush,");
-                sbValues.append(booleanToInt(player.isHasPreOrderArr()) + "," + booleanToInt(player.isHasPreOrderHW()) + ","
-                                + booleanToInt(player.isHasPreOrderSB()) + "," + booleanToInt(player.isHasARRArtbook()) + ","
-                                + booleanToInt(player.isHasHWArtbookOne()) + "," + booleanToInt(player.isHasHWArtbookTwo()) + ","
-                                + booleanToInt(player.isHasEncyclopediaEorzea()) + "," + booleanToInt(player.isHasBeforeMeteor()) + ","
-                                + booleanToInt(player.isHasBeforeTheFall()) + "," + booleanToInt(player.isHasSoundtrack()) + ","
-                                + booleanToInt(player.isHasAttendedEternalBond()) + "," + booleanToInt(player.isHasCompletedHWSightseeing())
-                                + "," + booleanToInt(player.isHasCompleted2pt5()) + "," + booleanToInt(player.isHasFiftyComms()) + ","
-                                + booleanToInt(player.isHasMooglePlush()) + "," + booleanToInt(player.isHasTopazCarbunclePlush()) + ","
-                                + booleanToInt(player.isHasEmeraldCarbunclePlush()) + ",");
-
-                sbFields.append("hildibrand, ps4collectors, dideternalbond, arrcollector, kobold, sahagin, amaljaa, "
-                                + "sylph, moogle, vanuvanu, vath, hw_complete, hw_31_complete, hw_33_complete, legacy_player");
-                sbValues.append(booleanToInt(player.isHasCompletedHildibrand()) + "," + booleanToInt(player.isHasPS4Collectors()) + ","
-                                + booleanToInt(player.isHasEternalBond()) + "," + booleanToInt(player.isHasARRCollectors()) + ","
-                                + booleanToInt(player.isHasKobold()) + "," + booleanToInt(player.isHasSahagin()) + ","
-                                + booleanToInt(player.isHasAmaljaa()) + "," + booleanToInt(player.isHasSylph()) + ","
-                                + booleanToInt(player.isHasMoogle()) + "," + booleanToInt(player.isHasVanuVanu()) + ","
-                                + booleanToInt(player.isHasVath()) + "," + booleanToInt(player.isHasCompletedHW()) + ","
-                                + booleanToInt(player.isHasCompleted3pt1()) + "," + booleanToInt(player.isHasCompleted3pt3()) + ","
-                                + booleanToInt(player.isHasCompletedSB()) + "," + booleanToInt(player.isLegacyPlayer()));
-
-            }
-
-            if(appConfig.isStoreMinions()) {
-                sbFields.append(",");
-                sbValues.append(",");
-                sbFields.append("minions");
-                sbValues.append("\"" + StringUtils.join(player.getMinions(), ",") + "\"");
-            }
-            if(appConfig.isStoreMounts()) {
-                sbFields.append(",");
-                sbValues.append(",");
-                sbFields.append("mounts");
-                sbValues.append("\"" + StringUtils.join(player.getMounts(), ",") + "\"");
-            }
-
-            if(appConfig.isStoreActiveDate()) {
-                sbFields.append(",");
-                sbValues.append(",");
-                sbFields.append("date_active");
-                java.text.SimpleDateFormat sdf = new java.text.SimpleDateFormat("yyyy-MM-dd");
-
-                String sqlDate = sdf.format(player.getDateImgLastModified());
-                sbValues.append("\"" + sqlDate + "\"");
-            }
-            if(appConfig.isStorePlayerActive()) {
-                sbFields.append(",");
-                sbValues.append(",");
-                sbFields.append("is_active");
-                sbValues.append(booleanToInt(player.isActive()));
-            }
-
-            sbFields.append(")");
-            sbValues.append(");");
-
-            String strSQL = sbFields.toString() + sbValues.toString();
-
-            st.executeUpdate(strSQL);
-            if(!appConfig.isQuiet() || isVerbose()) {
-                strOut = "Character " + player.getId() + " written to database successfully.";
-            }
-            closeConnection(conn);
-        } catch(SQLException e) {
-            e.printStackTrace();
-        }
-        return strOut;
-    }
-
-    private int booleanToInt(final boolean value) {
-        return BooleanUtils.toInteger(value);
     }
 
     /**
